added conditional rendering to MovieDetails, MovieDetails functional
